better explain Property <unknown>
I had thought it could happen when passing the wrong type to a Property
constructor, but happily that is not the case. For example

	& osDebian Unstable False

Still produces a good error:

    • Couldn't match expected type ‘Architecture’
                  with actual type ‘Bool’

So Property <unknown> only happens when too many or too few parameters
are passed, eg:

	& osDebian Unstable X86_64 False
	& osDebian Unstable

Before commit 14f6ae30809d8bbdb10b91cc59757e865a365df8,
the former of those resulted in:

    • Couldn't match expected type ‘[Char] -> p0’
                  with actual type ‘Property (HasInfo + Debian)’
    • The function ‘osDebian’ is applied to three arguments,
      but its type ‘DebianSuite
                    -> Architecture -> Property (HasInfo + Debian)’
      has only two

Unfortunately, it's now just Property <unknown>. This is
the only reversion I've identified from commit
14f6ae30809d8bbdb10b91cc59757e865a365df8.
My ghc bug https://gitlab.haskell.org/ghc/ghc/issues/16894
isn't quite explaining this problem yet, and I need to find a minimal
case of it.

As to the latter with too few parameters, it used to result in
a massive ugly useless error. Now it is also Property <unknown>,
which is an improvement.