use security.debian.org/debian-security
There's supposed to be a /debian-security path in the url, according to
various sources, including https://www.debian.org/security/

I tried it both ways for both stable and testing and oldstable, and it works
with and without it for both. Presumably some url rewriting is fixing it up on
the back end.