tail the dist-newstyle find
Unfortunately builds for previous versions can linger in there. Which
may be a cruft accumulation problem generally, but for now I don't want
it to fail if more than one is found.

Assuming that the last item in the find will be the newest.

This commit was sponsored by Brock Spratlen on Patreon.