stop using dist/setup-config as indication propellor is configured
cabal is changing to new-dist, and does not write setup-config in the
same place.

Instead, use a top-level "configured" stamp file. Same as the Makefile
does now.

This will lead to one extra run of cabal configure on each host, in
order to get the new stamp file written.

This commit was sponsored by Jochen Bartl on Patreon.