fix one more hardcoded dist/ path
Refactored the cabal exec code to reuse it here.

Had to change the location of propellor.built too, so put it in the top
of the propellor repo, rather than in dist.

This seems like it could be simpler. Why does propellor need to symlink
to propellor.built? Why not copy to a temp file, and atomically rename
it to propellor? It seems possible that something could depend on it
being a symlink, but I don't know what.

Adding to my confusion, there's the comment about cp -pfRL and timestamp
checking. What timestamp checking? The commit adding the comment didn't
change anything else, and I can't find any now or then.

This commit was sponsored by Jack Hill on Patreon.